feat: add hasFeatureFlagEnabled util function
